ci: run pipeline for Torizon samples
Add a build stage named "bridge" to trigger pipelines in other projects.
Add a job to trigger the samples pipeline under the "bridge" stage.
This makes sure that the demos still build successfully for new versions
of the Debian containers.

Allow the bridge-build-samples job to fail. This way, the logic for the
pipeline to succeed stays the same as before this commit, and the
current workflow for the debian-cross-toolchains development team stays
the same.

Related-to: DWC-2487

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>