Refactor chore job to fetch arm64/armv7 archs for the base debian image
This solves a bug where, for images in which we do not specify the platform,
it just pulls the debian:<tag> that is available in the registry, which
before this commit was wrongly pulled as "amd64".

The combination of not specifying arch when building plus not having the
correct arch being pulled in the chores job broke our shipping target.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>