graphics-tests-vivante: install mesa-utils package
We currently use the regular graphics-tests to run es2_info on iMX8
platforms. This is suboptimal given that we also pull
graphics-tests-vivante to run other tests on them.

This commit introduces the mesa-utils package on graphics-tests-vivante
which contains es2_info binary, meaning we can run tests with a single
pull for iMX8 devices.

Related-to: TCCP-708, TOR-3400
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>