gitlab-ci: update registry variable
For official deployments into our registry we use the torizoncore
user instead of just 'torizon'. I missed this in the previous commit,
as for users, the repository namespace and the username is the same.

This commit adjusts the variable so everything works for the official
release pipeline.

I've also changed the registry variable to the more general, technical
term 'namespace'.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>