am62: chromium: add default flag --disable-gpu-memory-buffer-video-frames
Chromium on am62 freezes and hang in maximized mode. This is an upstream
bug [1] and the known workaround is to pass the aforementioned flag to
chromium. It has a small performance penalty (~30% fps reduction on
aquarium webgl test), but enables us to have hw accelerated webGL.

[1] https://issues.chromium.org/issues/331796411

Related-to: TCCP-814
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>
Bugfix: Add --disable-gpu-memory-buffer-video-frames to Chromium's startup script on AM62