am62: chromium-tests: update chromium's arguments and don't launch headless
Make the tests use the same arguments as we have in
torizon/chromium-am62's entrypoint (start-chromium.sh). And also do not
launch it headless because it wasn't working with our combination of
puppeter, weston, wayland and chromium. Launching the browser make it
work as expected.

Related-to: TCCP-814
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>
Fix: am62: chromium-tests: update chromium's arguments and don't launch headless