debian-docker-images: support-files: unify all files used in our images
We had a problem of duplication of configs, scripts and support files
previously. Most of them were almost the same or exactly the same. This
created duplication of code, made changes in MRs/commits longer and/or
let some scripts rotten. To deal with it, this commit symlinks all those
files to support-files folder.

In order to accomplish this unification, more changes were needed in a
few places. Namely, gitlab-ci and upstream weston's dockerfile.

Changes on gitlab-ci:

- fix shellcheck/shfmt jobs after script symlinks were added;
- fix shellcheck jobs ignoring $SHELLCHECK_SEVERITY variable and
  remove shellcheck-git since we use find now instead of git;
- decouples Dockerfile path from build context path in order to share
  files;
- invoke make_feature_map.sh before running build-weston job so
  upstream container also has it and can make a decision whether to
  enable GPU.

Changes in upstream weston's dockerfile:

- add imx_features folder to image so entry.sh can decide whether to use
  or not the GPU.

Related-to: TCCP-755, TCCP-834
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>