Run make_feature_map for am62
So, not ideal because right now make_feature_map only contains
information specific to iMX8, but it's inevitable that we'll need it
for the upcoming boards and it's just a small file for the other boards.

Ideally, we'd embed some type of metadata like the feature map (but
possibly more general, for things like VPU, NPU, DPU...) into each
base Debian image for each platform.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>