build: base: am62, imx8: build for amd64 as well
This sounds a little weird but the amd64 images are not meant to
be run in the device. They are, however, to be used as a base for
images such as the cross-toolchain one: in such cases on needs to be
able to link against libraries build for arm64 but using an x86 linker.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>