Add a virtual multiarch image build
This is used for images that have different base images with different arches.

For example, this is the dependency for the Chromium image:

                     +---------------------+
                     |  Chromium Multiarch |
                     +---------------------+
                              |
           +------------------+------------------+
           |                                     |
           v                                     v
+----------------------|             +-----------------+
| wayland-base-vivante |             |   wayland-base  |
+----------------------|             +-----------------+
        (arm64)                             (armv7)

Meaning it's impossible to build it with buildx in "one-go", as the FROM
instruction has to be completely solved by the start of the build (ie, no
conditionals in FROM).

Related-to: TCCP-852

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>