any: add wayland-base-dev for upstream and platform-specific containers
wayland-base-dev comes with the most common build tools used to develop
software and a few graphical libraries. The main goal of it is to
provide a default base to use as a build stage for compiling graphical
applications. For instance, we change our test containers to use as base
for the build stage giving the benefit of declaring less packages to be
installed, improving the speed and making the Dockerfiles universal.

Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>