imx8: add QT Enterprise Demo container and clear screenshots' licensing
The container is multi-staged to only include the applications and
run-time dependencies in the final image, but there is a build stage
where the demo applications from QT are compiled. It requires a QT
license and account to build - username and password must be defined as
CI/CD variables. Tests make sure all aplications used in the demo are
able to run without errors.

Related-to: TCCP-967
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>