demo-gallery: add basic structure to generate the demo gallery
The demo gallery is a collection of docker compose files and some
attached metadata that can be used to build a delegation.

As Torizon Cloud can't understand the "platform" concept we have to
specify each machine, even though the compose file is the same for
all, for example, apalis-imx8 and verdin-imx8mm use the same compose
but they must be specified twice.

This means it's a pain to maintain these by hand, so we generate them
on the fly from the 'platform' compose files.

Also, these compose files must be canonicalized, which is done using tcb.

Related-to: TCCP-814

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>