tests: Improve logging of xray-junit-uploader-wrapper
Previously, the xray-junit-uploder logs did not make it clear which request it
was trying to send to JIRA.

This commit enhances the script by adding informative log messages that clarify
which report is being uploaded, which test execution key is being used, and
which test plan is targeted. This improves visibility during CI runs.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>