ci-scripts/test: Return error if test execution or test plan key is not set
Test execution and test plan keys variables are requirements for
"xray-junit-uploader-wrapper" script to be successful, however the script is
returning only a warning, which does not cause the job to fail in the CI. This
has resulted in forgetting to add new machines in Xray, since the test passing
gives the false impression that everything is fine.

This commit changes this behavior to return an error if this variable is
not set, which indicates better that something needs to be fixed, in
that case defining this variable.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>