ci-scripts/test: Run only one AVAL job for the same machine at the same time
Previously, multiple jobs could run using the same AVAL machine. However, in
practice, there is usually only one machine of each type in AVAL. As a result,
one of these jobs will run while the others wait to acquire the machine,
retrying and occupying runners, which can even result in a timeout if the
machine cannot be locked. This is the reason for the error
"Wasn't able to lock the device <uuid> after 80 attempts."

This commit makes use of "resource_group" feature of GitLab CI, ensuring that
only one job can run on a machine at a time, while the other jobs are put on
hold. This prevents unnecessary consumption of runner resources and avoids
timeouts caused by failing to lock the device. This mechanism even works
across different pipelines: only one job in the entire project can access the
resource at a time.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>