ci-scripts: use the "all" tag for the AVAL end-to-end tests
Previously, these jobs used the "torizoncore" tag, which is specific to a
legacy Horw runner with internal network access.

Since these jobs have no dependency on that specific runner and can run on any
runner with internal network access, this commit updates them to use the "all"
tag instead.

The "all" tag is available on all legacy Horw runners, making it more
appropriate and preventing overload on a single runner.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>